+++ kbe/src/lib/db_postgresql/db_interface_postgresql.h
@@ -54,7 +54,7 @@ public:
 	virtual bool detach();
 	bool reattach();
 
-	virtual bool query(const char* cmd, uint32 size, bool printlog = true, MemoryStream* result = NULL);
+	virtual bool query(const char* cmd, size_t size, bool printlog = true, MemoryStream* result = NULL);
 	bool write_query_result(PGresult* pgResult, MemoryStream* result);
 	PGresult* queryResult(const std::string& sql, const char* caller,
 		ExecStatusType expectedStatus = PGRES_TUPLES_OK, bool traceQuery = true);
