+++ kbe/src/lib/server/python_app.cpp
@@ -445,7 +445,7 @@ PyObject* PythonApp::__py_getAppPublish(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_setScriptLogType(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::scriptLogType(): args error!");
@@ -469,7 +469,7 @@ PyObject* PythonApp::__py_setScriptLogType(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_getResFullPath(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::getResFullPath(): args error!");
@@ -496,7 +496,7 @@ PyObject* PythonApp::__py_getResFullPath(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_hasRes(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::hasRes(): args error!");
@@ -519,7 +519,7 @@ PyObject* PythonApp::__py_hasRes(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_kbeOpen(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount != 2)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::open(): args error!");
@@ -560,7 +560,7 @@ PyObject* PythonApp::__py_kbeOpen(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_matchPath(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::matchPath(): args error!");
@@ -584,7 +584,7 @@ PyObject* PythonApp::__py_matchPath(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* PythonApp::__py_listPathRes(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	if(argCount < 1 || argCount > 2)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::listPathRes(): args[path, pathargs=\'*.*\'] error!");
