+++ kbe/src/server/tools/kbcmd/kbcmd.cpp
@@ -175,9 +175,9 @@ void KBCMD::finalise()
 //-------------------------------------------------------------------------------------
 int KBCMD::creatDir(const char *pDir)
 {
-	int i = 0;
+	size_t i = 0;
 	int iRet = -1;
-	int iLen = 0;
+	size_t iLen = 0;
 	char* pszDir = NULL;
 
 	if (NULL == pDir)
